REF example prog
