+++ src/aic_example/example_main.hpp
@@ -1,0 +1,3 @@
+
+void example_setup();
+void example_loop();
